requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.2
PyPDF2>=3.0.1
faiss-cpu>=1.9.0
sentence-transformers>=2.5.1
nltk>=3.8.1
numpy>=1.26.0,<2.0.0
pandas>=2.0.0
plotly>=5.18.0
python-dateutil>=2.8.2
tqdm>=4.67.0
protobuf>=4.25.0,<5.0.0
